+++ src/osmnet/build_graph.py
@@ -1,13 +1,9 @@
 import sys
-
 sys.path.append('..')
 
 from graph import GeoDigraph
-
-from db.db_process import gdf_to_postgis
 from osmnet.downloader import download_osm_xml
 from osmnet.parse_osm_xml import parse_xml_to_graph
-from osmnet.osm_io import load_graph, save_graph
 
 
 def load_geograph(ckpt):
@@ -34,13 +30,13 @@ def build_geograph(xml_fn=None, bbox=None, ckpt=None, way_info=True, *args, **kw
     
 
 if __name__ == "__main__":
-    graph = build_geograph("../../cache/GBA.osm.xml")
-    # graph.save_checkpoint('../../cache/GBA.graph.ckpt')
+    graph = build_geograph("../../cache/Shenzhen.osm.xml")
+    graph.save_checkpoint('../../cache/Shenzhen_graph_9.graph.ckpt')
     
     path = graph.search(src=7959990710, dst=499265789)
     graph.df_nodes.loc[path['path']].plot()
 
-    # save to DB
-    gdf_to_postgis(graph.df_edges.rename(columns={'highway': 'road_type'}), 'topo_osm_shenzhen_edge')
-    gdf_to_postgis(graph.df_nodes, 'topo_osm_shenzhen_node')
+    # # save to DB
+    # gdf_to_postgis(graph.df_edges.rename(columns={'highway': 'road_type'}), 'topo_osm_shenzhen_edge')
+    # gdf_to_postgis(graph.df_nodes, 'topo_osm_shenzhen_node')